Split ICDEWinForm to a file.
Changing Size field in CommonEntry to a long it is just as thats the type from OS anyway.
Added human readable size of a Folder to list views.
Created a helper for drop downs with Display and Value fields.
Created a helper for checkbox with dependent controls being enabled.
Modify DirEntry.SizeCompareWithDirTo() so that the size of the directory which is held in data structure is kept as well.
Implemented the advanced search controls.
Created MyMessageBox that centers on parent to avoid MessageBox.Show().

+++ src/cdeLibTest/DuplicationTest.cs
@@ -141,9 +141,9 @@ namespace cdeLibTest
             Console.WriteLine("Number of Dupe Pairs {0}", dupePairEnum.Count);
         }
 
-        private static ulong GetSumOfUniqueHashesForEachSize_ExcludePartialHash(IEnumerable<KeyValuePair<ulong, List<PairDirEntry>>> sizePairDictionary)
+        private static long GetSumOfUniqueHashesForEachSize_ExcludePartialHash(IEnumerable<KeyValuePair<long, List<PairDirEntry>>> sizePairDictionary)
         {
-            var sumOfUniqueHashesForEachSize = 0ul;
+            var sumOfUniqueHashesForEachSize = 0L;
             foreach (var list in sizePairDictionary)
             {
                 var fdeListOfSize = list.Value;
@@ -164,7 +164,7 @@ namespace cdeLibTest
                         }
                     }
                 }
-                sumOfUniqueHashesForEachSize = sumOfUniqueHashesForEachSize + (ulong)seenHash.Count;
+                sumOfUniqueHashesForEachSize = sumOfUniqueHashesForEachSize + seenHash.Count;
             }
             return sumOfUniqueHashesForEachSize;
         }
